chore(quality): move types into *.types.ts files, enforce types-location lane
