+++ .github/quality-gate/enforced/types-location
@@ -1,0 +1,1 @@
+enforced
