+++ apps/web/src/features/landing/components/demo/todos-demo/todosDemoConstants.ts
@@ -1,36 +1,11 @@
-export interface DemoProject {
-  id: string;
-  name: string;
-  color: string;
-}
+import type { DemoProject, DemoTodo } from "./todosDemoConstants.types";
 
-export interface DemoSubTask {
-  id: string;
-  title: string;
-  completed: boolean;
-}
-
-export interface DemoWorkflowStep {
-  id: string;
-  title: string;
-  description: string;
-  category: string;
-}
-
-export interface DemoTodo {
-  id: string;
-  title: string;
-  description?: string;
-  priority: "high" | "medium" | "low" | "none";
-  labels: string[];
-  project_id: string;
-  due_date?: string;
-  completed: boolean;
-  subtasks: DemoSubTask[];
-  workflow_categories: string[];
-  workflow_steps: DemoWorkflowStep[];
-  created_at: string;
-}
+export type {
+  DemoProject,
+  DemoSubTask,
+  DemoTodo,
+  DemoWorkflowStep,
+} from "./todosDemoConstants.types";
 
 export const DEMO_PROJECTS: DemoProject[] = [
   { id: "inbox", name: "Inbox", color: "#71717a" },
