+++ apps/web/src/features/landing/data/featuresData.ts
@@ -3,51 +3,16 @@ import { AUTOMATION_FEATURES } from "./features/automation";
 import { INTEGRATIONS_FEATURES } from "./features/integrations";
 import { MULTI_PLATFORM_FEATURES } from "./features/multiPlatform";
 import { PRODUCTIVITY_FEATURES } from "./features/productivity";
+import type { FeatureCategory, FeatureData } from "./featuresData.types";
 
-export type FeatureCategory =
-  | "AI Intelligence"
-  | "Productivity"
-  | "Automation"
-  | "Integrations"
-  | "Multi-Platform";
-
-export interface FeatureBenefit {
-  icon: string;
-  title: string;
-  description: string;
-}
-
-export interface HowItWorksStep {
-  number: string;
-  title: string;
-  description: string;
-}
-
-export interface FeatureFAQ {
-  question: string;
-  answer: string;
-}
-
-export interface FeatureUseCase {
-  title: string;
-  description: string;
-}
-
-export interface FeatureData {
-  slug: string;
-  category: FeatureCategory;
-  icon: string;
-  title: string;
-  tagline: string;
-  headline: string;
-  subheadline: string;
-  benefits: [FeatureBenefit, FeatureBenefit, FeatureBenefit];
-  howItWorks?: [HowItWorksStep, HowItWorksStep, HowItWorksStep];
-  faqs?: [FeatureFAQ, FeatureFAQ, FeatureFAQ, FeatureFAQ];
-  useCases?: [FeatureUseCase, FeatureUseCase, FeatureUseCase];
-  relatedSlugs?: [string, string, string];
-  demoComponent: string;
-}
+export type {
+  FeatureBenefit,
+  FeatureCategory,
+  FeatureData,
+  FeatureFAQ,
+  FeatureUseCase,
+  HowItWorksStep,
+} from "./featuresData.types";
 
 export const FEATURE_CATEGORIES: FeatureCategory[] = [
   "AI Intelligence",
