+++ packages/cli/src/lib/env-parser.ts
@@ -1,35 +1,19 @@
 import * as fs from "node:fs";
 import * as path from "node:path";
 import { execa } from "execa";
+import type {
+  EnvCategory,
+  EnvVar,
+  SetupMode,
+  WebEnvVar,
+} from "./env-parser.types";
 
-export type SetupMode = "selfhost" | "developer";
-
-export interface EnvVar {
-  name: string;
-  required: boolean;
-  category: string;
-  description: string;
-  affectedFeatures: string;
-  defaultValue?: string;
-  docsUrl?: string;
-}
-
-export interface EnvCategory {
-  name: string;
-  description: string;
-  affectedFeatures: string;
-  requiredInProd: boolean;
-  allRequired: boolean;
-  docsUrl?: string;
-  alternativeGroup?: string;
-  variables: EnvVar[];
-}
-
-export interface WebEnvVar {
-  name: string;
-  value: string;
-  category: string;
-}
+export type {
+  EnvCategory,
+  EnvVar,
+  SetupMode,
+  WebEnvVar,
+} from "./env-parser.types";
 
 // Infrastructure connection strings set by setup mode.
 //
